+++ modules/ui_extra_networks_checkpoints.py
@@ -63,13 +63,8 @@ class ExtraNetworksPageCheckpoints(ui_extra_networks.ExtraNetworksPage):
             }
             record["info"] = self.find_info(checkpoint.filename)
             record["description"] = self.find_description(checkpoint.filename, record["info"])
-            
-            # Base model extraction
-            record["version"] = (
-                record["info"].get("modelVersions", [{}])[0].get("baseModel", "N/A")
-                if record["info"] else "N/A"
-            )
-            
+            record["version"] = record["info"].get("modelVersions", [{}])[0].get("baseModel", "N/A") if record["info"] else "N/A"
+
         except Exception as e:
             shared.log.debug(f'Networks error: type=model file="{name}" {e}')
         return record