+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   - update `requirements`  
 - **UI**  
   - improved image scaling in img2img and control interfaces  
+  - add base model type to networks display, thanks @Artheriax  
 - **Fixes**
   - normalize path hanlding when deleting images  
   - fix hidden model tags in networks display